PWA-2159 Moved some components out of the filter page into the library.
