+++ libraries/engage/components/RangeSlider/style.js
@@ -1,5 +1,9 @@
 import { css } from 'glamor';
 
+const container = css({
+  position: 'relative',
+});
+
 const outerRange = css({
   minHeight: 1,
   position: 'relative',
@@ -11,20 +15,8 @@ const range = css({
   position: 'absolute',
 }).toString();
 
-const srOnly = css({
-  position: 'absolute',
-  width: 1,
-  height: 1,
-  padding: 0,
-  margin: -1,
-  overflow: 'hidden',
-  clip: 'rect(0, 0, 0, 0)',
-  whiteSpace: 'nowrap',
-  border: 0,
-});
-
 export default {
+  container,
   outerRange,
   range,
-  srOnly,
 };
