+++ libraries/engage/filter/components/FilterItem/index.jsx
@@ -3,16 +3,18 @@ import PropTypes from 'prop-types';
 import styles from './style';
 
 /**
+ * Wraps a single fiter page item.
+ * @param {Object} props The component props.
  * @returns {JSX}
  */
-const FilterContentItem = ({ children }) => (
+const FilterItem = ({ children }) => (
   <div className={styles}>
     {children}
   </div>
 );
 
-FilterContentItem.propTypes = {
+FilterItem.propTypes = {
   children: PropTypes.node.isRequired,
 };
 
-export default FilterContentItem;
+export default FilterItem;
