+++ libraries/engage/filter/components/PriceSlider/style.js
