+++ libraries/engage/filter/index.js
@@ -13,3 +13,12 @@ export * from '@shopgate/pwa-common-commerce/filter/selectors';
 
 // STREAMS
 export * from '@shopgate/pwa-common-commerce/filter/streams';
+
+// COMPONENTS
+export { default as FilterItem } from './components/FilterItem';
+export { default as PriceSlider } from './components/PriceSlider';
+
+// HELPERS
+
+export { default as buildInitialFilters } from './helpers/buildInitialFilters';
+export { default as buildUpdatedFilters } from './helpers/buildUpdatedFilters';
