--- themes/theme-ios11/pages/Filter/components/Content/components/Item/index.jsx
@@ -1,18 +1,0 @@
-import React from 'react';
-import PropTypes from 'prop-types';
-import styles from './style';
-
-/**
- * @returns {JSX}
- */
-const FilterContentItem = ({ children }) => (
-  <div className={styles}>
-    {children}
-  </div>
-);
-
-FilterContentItem.propTypes = {
-  children: PropTypes.node.isRequired,
-};
-
-export default FilterContentItem;
