--- themes/theme-ios11/pages/Filter/components/Content/components/Item/style.js
@@ -1,9 +1,0 @@
-import { css } from 'glamor';
-import { themeConfig } from '@shopgate/pwa-common/helpers/config';
-
-const { colors } = themeConfig;
-
-export default css({
-  background: colors.light,
-  marginTop: 4,
-});
