--- themes/theme-ios11/pages/Filter/components/Content/helpers/buildInitialFilters.js
@@ -1,32 +1,0 @@
-import { FILTER_TYPE_RANGE } from '@shopgate/pwa-common-commerce/filter/constants';
-
-/**
- * @param {Object} filters The available filters.
- * @param {Object} activeFilters The selected filters.
- * @returns {Object}
- */
-const buildInitialFilters = (filters, activeFilters) => {
-  const defaults = {};
-
-  if (!filters) {
-    return defaults;
-  }
-
-  filters.forEach((filter) => {
-    const {
-      id, maximum, minimum, type,
-    } = filter;
-
-    defaults[id] = {
-      type,
-      value: (type === FILTER_TYPE_RANGE) ? [minimum, maximum] : [],
-    };
-  });
-
-  return {
-    ...defaults,
-    ...activeFilters,
-  };
-};
-
-export default buildInitialFilters;
